split/reuse eth_tester
